Refactor: Split PocketBase services into modular structure
This major refactoring splits the monolithic pocketbase.ts (~5000 lines) into
a well-organized modular structure for improved maintainability.

Structure:
- src/services/pocketbase/
  - client.ts - PocketBase client initialization
  - context.ts - Site context management
  - types/ - All type definitions (14 files)
  - services/ - All service classes (22 files)
  - index.ts - Module barrel export

Key improvements:
- Clear separation of concerns with focused service files
- Shared mapper utilities to avoid code duplication
- Dependency injection for circular dependencies
- Full backwards compatibility via re-exports
- Includes new AnalyticsSettingService from main

All 3857 tests pass with type checking clean.
