+++ src/services/pocketbase/client.ts
@@ -1,0 +1,14 @@
+/**
+ * PocketBase Client
+ * Single instance of PocketBase client for the entire application
+ */
+
+import PocketBase from 'pocketbase';
+
+// Get PocketBase URL from environment variables with fallback
+const POCKETBASE_URL = import.meta.env.VITE_POCKETBASE_URL || 'http://localhost:8090' || 'http://127.0.0.1:8090';
+
+export const pb = new PocketBase(POCKETBASE_URL);
+
+// Enable auto cancellation for duplicate requests
+pb.autoCancellation(true);
